requirements.txt
openai
python-dotenv
wikipedia
tiktoken
neo4j
streamlit
sentence_transformers==2.2.2
Pillow
fastapi
PyPDF2
torch==2.0.1
pydantic
uvicorn
sse-starlette
boto3
# missing from the langchain base image?
langchain-openai
langchain-community
langchain-google-genai
selenium==4.7.2 
transformers==4.25.1 
graphdatascience
xvfbwrapper
xmltodict
praw
trafilatura
lxml_html_clean
google-cloud-storage
youtube-transcript-api
